:truck: move indexing logic into its own usecase
