+++ marketplace-indexer/src/application/mod.rs
@@ -1,2 +1,5 @@
 mod builders;
 pub use builders::*;
+
+mod indexing;
+pub use indexing::Usecase as IndexingUsecase;
